[IMP] calendar: show more meeting details
Before this commit there is no label shown at the top of the invite if
the user hasn't done anything with the invite or if he chose
'Tentative'. Now it always shows the informative label. This commit
also adds the description of a meeting subject as it can be
interesting and important in many cases.

closes odoo/odoo#43121

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>