[IMP] base: do not load actions when can not read target models
As neither fields_view_get, nor default_get actually verify you have
access to a specific model (this is likely to be changed in the
future), it was possible to execute an action to a model, even if the
user does not have access to the corresponding model.
Before this commit, launching an action to a wizard with no access was
opening a new record view with all fields in readonly and clicking on
any button raised an access error.

To avoid the frustrating of seeing an action but getting an unusable
screen when clicking on it, just hide the action in the first place.