[FIX] point_of_sale: payment method name not displayed
When a payment line is created, and then the order is reloaded, by the
synchronization feature for example, the name displayed on the payment
line is empty.

This happens because the field name is not saved on the server, so
instead of using it, we are directly taking the name of the payment
method linked to the payment line.

closes odoo/odoo#43615

X-original-commit: 006371ada2bd8c46a2d37775bf30ae1ac30974ec
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>