[FIX] mail: restores the documentViewer modal
Revision on
https://github.com/odoo/odoo/commit/f661077522cf7de265c1df91b39cf2e63da47017#diff-9da02f8f30ea349fbc4b9884c2b77b78

The document viewer did not work properly because, by default, dialogs are non-active,
so most interactions were not possible (just pressing ESC to close the dialog worked).

The correct fix consists of using OwlDialog for DocumentViewer,
but that's basically close to a re-writing, so in the meantime marking it as always active is fine.

closes odoo/odoo#43446

Task: 2177011
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>