[IMP] account_check_printing: log state for vendor checks
Task 2168197

Especially when writing out checks we would like to track the value of
the state of the vendor payment in the chatter. This way we can see who
prints it, cancels it, etc and at what time this happened. It's
important for traceability. I think we can just enable it for all
payments in general, I can't think of any downsides and payments are
usually sensitive documents so I imagine people want as much info as
possible.

closes odoo/odoo#43505

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>